move module payment screen
