+++ ui/src/main/java/com/jin/ui/payment/PaymentDetailScreen.kt
@@ -1,4 +1,4 @@
-package com.jin.honey.feature.paymentdetail.ui
+package com.jin.ui.payment
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
@@ -17,12 +17,12 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
 import com.jin.feature.ui.state.UiState
-import com.jin.honey.R
-import com.jin.honey.feature.paymentdetail.ui.content.PayDetailInformation
-import com.jin.honey.feature.paymentdetail.ui.content.PayDetailOrderContent
-import com.jin.honey.feature.paymentdetail.ui.content.PayDetailOrderInfo
-import com.jin.honey.feature.paymentdetail.ui.content.PayDetailOrderPrice
-import com.jin.honey.feature.paymentdetail.ui.content.PayDetailOverView
+import com.jin.ui.R
+import com.jin.ui.payment.content.PayDetailInformation
+import com.jin.ui.payment.content.PayDetailOrderContent
+import com.jin.ui.payment.content.PayDetailOrderInfo
+import com.jin.ui.payment.content.PayDetailOrderPrice
+import com.jin.ui.payment.content.PayDetailOverView
 import com.jin.domain.cart.model.Cart
 
 @Composable
