+++ ui/src/main/java/com/jin/ui/payment/content/PayDetailInformation.kt
@@ -1,4 +1,4 @@
-package com.jin.honey.feature.paymentdetail.ui.content
+package com.jin.ui.payment.content
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -11,7 +11,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.jin.honey.R
+import com.jin.ui.R
 import java.time.Instant
 import java.time.ZoneId
 import java.time.format.DateTimeFormatter
